Refactor, default change
Change default from 1 month to 1 day for default cleanup options
csv_export model
 - remove hack for filters identified object state issue
 - move defaults to migration
 - removed nil from valid status
 - refactor scope old and delete_file method
Refactor csv_export_job,
 - rename export_task to generate_csv
 - remove notify_of_creation method
csv_mailer
 - refactor remove _email
 - improve email message contents
models/csv_export_test
 - add a describe block around status test blocks
 - remove unnecessary defaults tests

+++ app/models/csv_export.rb
@@ -1,9 +1,8 @@
 class CsvExport < ActiveRecord::Base
   belongs_to :user
   serialize :filters, JSON
-  STATUS_VALUES = [nil, 'pending', 'started', 'finished', 'downloaded', 'failed', 'deleted']
+  STATUS_VALUES = ['pending', 'started', 'finished', 'downloaded', 'failed', 'deleted']
 
-  before_create :set_defaults
   before_destroy :delete_file
 
   validates :status, inclusion: { in: STATUS_VALUES }
@@ -11,10 +10,10 @@ class CsvExport < ActiveRecord::Base
 
 
   scope :old, lambda {
-    end_date = Time.now - Rails.application.config.samson.export_job.downloaded_age
-    timeout_date = Time.now - Rails.application.config.samson.export_job.max_age
+    end_date = Rails.application.config.samson.export_job.downloaded_age.seconds.ago
+    timeout_date = Rails.application.config.samson.export_job.max_age.seconds.ago
     where("(status = 'downloaded' AND updated_at <= :end_date) OR updated_at <= :timeout_date",
-    {end_date: end_date, timeout_date: timeout_date})
+      end_date: end_date, timeout_date: timeout_date)
   }
 
   def status?(state)
@@ -30,8 +29,8 @@ class CsvExport < ActiveRecord::Base
   end
 
   def filters
-    filter = super || {}
-    if filter['deploys.created_at'].is_a?(String)
+    filter = super.clone
+    if filter['deploys.created_at']
       dates = filter['deploys.created_at'].scan(/-?\d+-\d+-\d+/)
       filter['deploys.created_at'] = (Date.parse(dates[0])..Date.parse(dates[1]))
     end
@@ -43,20 +42,6 @@ class CsvExport < ActiveRecord::Base
   end
 
   def delete_file
-    if File.exist?(path_file)
-      File.delete(path_file)
-      if File.exist?(path_file)
-        Rails.logger.error "Failed to delete file #{path_file}"
-      else
-        Rails.logger.info "Successfully deleted file #{path_file}"
-      end
-    end
-  end
-
-  private
-
-  def set_defaults
-    self.status ||= :pending
-    self.filters ||= {}
+    File.delete(path_file) if File.exist?(path_file)
   end
 end
